refactor folders and cleanup
