+++ crop_ocr_signs/connection.py
