+++ crop_ocr_signs/verify_signs/no_partial_order_x1.txt
