+++ crop_ocr_signs/verify_signs/remove_wrong_crops.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+
+folderA="$1"
+folderB="$2"
+
+for file in "$folderA"/*; do
+  filename=$(basename "$file")
+  if [[ ! -e "$folderB/$filename" ]]; then
+    echo "Removing $file (not in $folderB)"
+    rm -rf "$file"
+  fi
+done
