--- utils/stats/no_partial_order_percentages.json
@@ -1,120 +1,0 @@
-{
-  "ABZ1": { "proportion": 0.177, "total": 6327 },
-  "ABZ5": { "proportion": 0.166, "total": 1321 },
-  "ABZ6": { "proportion": 0.138, "total": 232 },
-  "ABZ7": { "proportion": 0.176, "total": 1470 },
-  "ABZ9": { "proportion": 0.467, "total": 30 },
-  "ABZ12": { "proportion": 0.27, "total": 226 },
-  "ABZ13": { "proportion": 0.157, "total": 19830 },
-  "ABZ15": { "proportion": 0.166, "total": 3047 },
-  "ABZ38": { "proportion": 0.235, "total": 187 },
-  "ABZ52": { "proportion": 0.213, "total": 906 },
-  "ABZ55": { "proportion": 0.17, "total": 3185 },
-  "ABZ58": { "proportion": 0.188, "total": 865 },
-  "ABZ59": { "proportion": 0.188, "total": 711 },
-  "ABZ60": { "proportion": 0.181, "total": 326 },
-  "ABZ61": { "proportion": 0.159, "total": 6417 },
-  "ABZ62": { "proportion": 0.175, "total": 80 },
-  "ABZ68": { "proportion": 0.173, "total": 2599 },
-  "ABZ69": { "proportion": 0.174, "total": 3674 },
-  "ABZ70": { "proportion": 0.167, "total": 7777 },
-  "ABZ72": { "proportion": 0.17, "total": 147 },
-  "ABZ73": { "proportion": 0.165, "total": 2624 },
-  "ABZ74": { "proportion": 0.152, "total": 394 },
-  "ABZ75": { "proportion": 0.193, "total": 4823 },
-  "ABZ78": { "proportion": 0.181, "total": 448 },
-  "ABZ79": { "proportion": 0.125, "total": 248 },
-  "ABZ80": { "proportion": 0.176, "total": 942 },
-  "ABZ84": { "proportion": 0.18, "total": 735 },
-  "ABZ85": { "proportion": 0.165, "total": 442 },
-  "ABZ86": { "proportion": 0.163, "total": 3949 },
-  "ABZ94": { "proportion": 0.333, "total": 6 },
-  "ABZ97": { "proportion": 0.151, "total": 755 },
-  "ABZ99": { "proportion": 0.185, "total": 1162 },
-  "ABZ101": { "proportion": 0.163, "total": 258 },
-  "ABZ104": { "proportion": 0.167, "total": 240 },
-  "ABZ111": { "proportion": 0.165, "total": 115 },
-  "ABZ112": { "proportion": 0.156, "total": 944 },
-  "ABZ115": { "proportion": 0.172, "total": 478 },
-  "ABZ128": { "proportion": 0.168, "total": 488 },
-  "ABZ134": { "proportion": 1.0, "total": 2 },
-  "ABZ139": { "proportion": 0.161, "total": 2980 },
-  "ABZ142": { "proportion": 0.188, "total": 4018 },
-  "ABZ143": { "proportion": 0.203, "total": 64 },
-  "ABZ144": { "proportion": 0.166, "total": 429 },
-  "ABZ145": { "proportion": 0.252, "total": 619 },
-  "ABZ147": { "proportion": 0.13, "total": 246 },
-  "ABZ151": { "proportion": 0.153, "total": 1529 },
-  "ABZ167": { "proportion": 0.193, "total": 327 },
-  "ABZ170": { "proportion": 0.25, "total": 44 },
-  "ABZ172": { "proportion": 0.194, "total": 619 },
-  "ABZ191": { "proportion": 0.244, "total": 160 },
-  "ABZ206": { "proportion": 0.195, "total": 3714 },
-  "ABZ207": { "proportion": 0.193, "total": 1017 },
-  "ABZ211": { "proportion": 0.204, "total": 372 },
-  "ABZ212": { "proportion": 0.116, "total": 199 },
-  "ABZ214": { "proportion": 0.174, "total": 4293 },
-  "ABZ230": { "proportion": 0.282, "total": 248 },
-  "ABZ231": { "proportion": 0.177, "total": 4702 },
-  "ABZ232": { "proportion": 0.246, "total": 65 },
-  "ABZ279": { "proportion": 0.18, "total": 1142 },
-  "ABZ295": { "proportion": 0.204, "total": 1347 },
-  "ABZ296": { "proportion": 0.168, "total": 1982 },
-  "ABZ306": { "proportion": 0.203, "total": 261 },
-  "ABZ308": { "proportion": 0.186, "total": 2959 },
-  "ABZ312": { "proportion": 0.211, "total": 350 },
-  "ABZ313": { "proportion": 0.194, "total": 315 },
-  "ABZ314": { "proportion": 0.171, "total": 416 },
-  "ABZ318": { "proportion": 0.185, "total": 6175 },
-  "ABZ319": { "proportion": 0.182, "total": 698 },
-  "ABZ322": { "proportion": 0.186, "total": 564 },
-  "ABZ324": { "proportion": 0.156, "total": 1747 },
-  "ABZ328": { "proportion": 0.166, "total": 2266 },
-  "ABZ330": { "proportion": 0.142, "total": 1628 },
-  "ABZ331": { "proportion": 0.213, "total": 296 },
-  "ABZ331e+152i": { "proportion": 0.195, "total": 149 },
-  "ABZ334": { "proportion": 0.193, "total": 709 },
-  "ABZ335": { "proportion": 0.152, "total": 961 },
-  "ABZ339": { "proportion": 0.282, "total": 39 },
-  "ABZ342": { "proportion": 0.174, "total": 11277 },
-  "ABZ343": { "proportion": 0.179, "total": 661 },
-  "ABZ353": { "proportion": 0.151, "total": 218 },
-  "ABZ354": { "proportion": 0.177, "total": 3533 },
-  "ABZ366": { "proportion": 0.179, "total": 3299 },
-  "ABZ367": { "proportion": 0.183, "total": 262 },
-  "ABZ371": { "proportion": 0.212, "total": 2270 },
-  "ABZ376": { "proportion": 0.162, "total": 790 },
-  "ABZ381": { "proportion": 0.172, "total": 7158 },
-  "ABZ383": { "proportion": 0.168, "total": 345 },
-  "ABZ384": { "proportion": 0.169, "total": 982 },
-  "ABZ396": { "proportion": 0.205, "total": 1361 },
-  "ABZ397": { "proportion": 0.153, "total": 118 },
-  "ABZ399": { "proportion": 0.153, "total": 570 },
-  "ABZ401": { "proportion": 0.208, "total": 424 },
-  "ABZ411": { "proportion": 0.179, "total": 1975 },
-  "ABZ427": { "proportion": 0.17, "total": 1500 },
-  "ABZ437": { "proportion": 0.152, "total": 66 },
-  "ABZ440": { "proportion": 0.265, "total": 34 },
-  "ABZ449": { "proportion": 0.156, "total": 5658 },
-  "ABZ457": { "proportion": 0.176, "total": 1578 },
-  "ABZ461": { "proportion": 0.171, "total": 6498 },
-  "ABZ465": { "proportion": 0.119, "total": 185 },
-  "ABZ468": { "proportion": 0.253, "total": 95 },
-  "ABZ470": { "proportion": 0.178, "total": 573 },
-  "ABZ480": { "proportion": 0.156, "total": 12220 },
-  "ABZ532": { "proportion": 0.18, "total": 1038 },
-  "ABZ533": { "proportion": 0.18, "total": 5617 },
-  "ABZ535": { "proportion": 0.209, "total": 445 },
-  "ABZ536": { "proportion": 0.155, "total": 1346 },
-  "ABZ537": { "proportion": 0.184, "total": 1247 },
-  "ABZ545": { "proportion": 0.16, "total": 5231 },
-  "ABZ554": { "proportion": 0.209, "total": 273 },
-  "ABZ570": { "proportion": 0.204, "total": 754 },
-  "ABZ575": { "proportion": 0.195, "total": 493 },
-  "ABZ579": { "proportion": 0.152, "total": 17082 },
-  "ABZ586": { "proportion": 0.158, "total": 450 },
-  "ABZ589": { "proportion": 0.162, "total": 758 },
-  "ABZ595": { "proportion": 0.232, "total": 164 },
-  "ABZ597": { "proportion": 0.177, "total": 12216 },
-  "ABZ598a": { "proportion": 0.1, "total": 10 }
-}
